- Moved to the presenter of the tab generale handling what tessere to display instead of the presenter of new mezzo to avoid code duplication - Server now provide some default data - Tab generale presenter can now display tessere currently linked or all tessere available if editing

+++ Flotta.ClientSide/TabGeneralePresenter.cs
@@ -14,29 +14,76 @@ namespace Flotta.ClientSide
 	{
 
 		private IServer _server;
-		private MezzoTabPresenter _tabs;
 		private ITabGeneraleView _view;
+
+		private MezzoTabPresenter _tabs;
+		private IMezzo _mezzo;
 
 		private IMezzo Mezzo
 		{
-			get => _tabs.Mezzo;
+			get => _tabs?.Mezzo ?? _mezzo;
 		}
 
 		private bool _editMode;
-		private bool EditMode
+		internal bool EditMode
 		{
 			get => _editMode;
 			set
 			{
 				_editMode = value;
+
+				if (value)
+				{
+					_tessereTmp = (from t in Mezzo.Tessere select t.Clone() as ITessera).ToList();
+					_tessereTypes = (from t in _server.TesseraTypes where !t.IsDisabled || (from tt in Mezzo.Tessere select tt.Type).Contains(t) select t).ToList();
+
+					//_dispositivoTmp = (from d in Mezzo.Dispositivi select d.Clone()).ToList();
+
+					//_permessiTmp = (from p in Mezzo.Permessi select p.Clone()).ToList();
+				}
+				else
+				{
+					_tessereTmp = null;
+					_tessereTypes = null;
+					_tessereItems = null;
+
+					_dispositivoTmp = null;
+
+					_permessiTmp = null;
+				}
+
 				Reload();
 			}
 		}
 
+		///<summary>During editing keeps a local copy of tessere linked to mezzo to edit freely.</summary>
+		private List<ITessera> _tessereTmp = null;
+		///<summary>During editing keeps the subset of tessere types available, aka non disabled or already linked to mezzo.</summary>
+		private List<ITesseraType> _tessereTypes = null;
+		///<summary>During editing keeps visual items for tessere corresponding to the list of available types _tessereTypes.</summary>
+		private List<ITesseraListItem> _tessereItems = null;
+
+		private List<IDispositivo> _dispositivoTmp = null;
+		// IDispositivoListItem
+
+		private List<IPermesso> _permessiTmp = null;
+		// IPermessoListItem
+
+		internal TabGeneralePresenter(IServer server, IMezzo mezzo, ITabGeneraleView view)
+		{
+			_mezzo = mezzo;
+			CommonContructor(server, view);
+		}
+
 		internal TabGeneralePresenter(IServer server, MezzoTabPresenter tabs, ITabGeneraleView view)
 		{
-			_server = server;
 			_tabs = tabs;
+			CommonContructor(server, view);
+		}
+
+		private void CommonContructor(IServer server, ITabGeneraleView view)
+		{
+			_server = server;
 			_view = view;
 
 			EditMode = false;
@@ -49,7 +96,7 @@ namespace Flotta.ClientSide
 
 		public void Reload()
 		{
-			if (_tabs.Mezzo == null)
+			if (Mezzo == null)
 				return;
 
 			_view.Modello = Mezzo.Modello;
@@ -63,12 +110,27 @@ namespace Flotta.ClientSide
 			_view.Profondita = Mezzo.Profondita;
 			_view.VolumeCarico = Mezzo.VolumeCarico;
 
+			if (_editMode)
+			{
+				_tessereItems = new List<ITesseraListItem>();
+				foreach (var tt in _tessereTypes)
+				{
+					ITessera tess = (from t in _tessereTmp where t.Type == tt select t).ElementAtOrDefault(0);
+					_tessereItems.Add(ClientSideInterfaceFactory.NewTesseraListItem(tess != null, tt.Name, tess?.Codice ?? "", tess?.Pin ?? ""));
+				}
+				_view.Tessere = _tessereItems;
+			}
+			else
+			{
+				_view.Tessere = from t in Mezzo.Tessere orderby t.Type.Name select ClientSideInterfaceFactory.NewTesseraListItem(true, t.Type.Name, t.Codice, t.Pin);
+			}
+
 			_view.EditMode = _editMode;
 		}
 
 		private void OnEnterEdit()
 		{
-			if (EditMode)
+			if (EditMode || _tabs == null)
 				return;
 
 			EditMode = true;
@@ -76,27 +138,32 @@ namespace Flotta.ClientSide
 
 		public void OnCancelEdit()
 		{
-			if (!EditMode)
+			if (!EditMode || _tabs == null)
 				return;
 
 			EditMode = false;
 		}
 
+		internal event GenericAction MezzoSaved;
 		private void OnSaveEdit()
 		{
 			if (!EditMode)
 				return;
 
-			var errors =_server.UpdateMezzo(_tabs.Mezzo, _view.Modello, _view.Targa, _view.Numero, _view.NumeroTelaio, _view.AnnoImmatricolazione, _view.Portata, _view.Altezza, _view.Lunghezza, _view.Profondita, _view.VolumeCarico, _tabs.Mezzo.Tessere, _tabs.Mezzo.Dispositivi, _tabs.Mezzo.Permessi);
+			var errors = _server.UpdateMezzo(_tabs.Mezzo, _view.Modello, _view.Targa, _view.Numero, _view.NumeroTelaio, _view.AnnoImmatricolazione, _view.Portata, _view.Altezza, _view.Lunghezza, _view.Profondita, _view.VolumeCarico, Mezzo.Tessere, Mezzo.Dispositivi, Mezzo.Permessi);
 
 			if (errors.Count() > 0) MessageBox.Show(String.Join("\r\n", errors), "Errore");
-			
+
 			// The tab will be automatically reloaded exiting edit mode automatically with the notification from the server
+			MezzoSaved?.Invoke();
 		}
 
 		private void OnDeleteMezzo()
 		{
-			if(MessageBox.Show("Sei sicuro di voler cancellare il mezzo corrente?", "Elimina mezzo", MessageBoxButtons.YesNo) == DialogResult.Yes)
+			if (_tabs == null)
+				return;
+
+			if (MessageBox.Show("Sei sicuro di voler cancellare il mezzo corrente?", "Elimina mezzo", MessageBoxButtons.YesNo) == DialogResult.Yes)
 			{
 				if (!_server.DeleteMezzo(_tabs.Mezzo))
 					MessageBox.Show("Errore durante l'eliminazione del mezzo", "Errore");
